requirements.txt
# Core
streamlit
openai
python-dotenv
requests

# Crawling
crawl4ai

# Vector DB
pinecone

# Embeddings
sentence-transformers

# Data processing
beautifulsoup4
html5lib
nltk

# Plotting (optional)
matplotlib
plotly

# Development
tqdm
